pull data from rethinkdb
I've been writing this entirely on Glitch so I haven't committed in a while, but here is what I have so far. You can see my progress on Glitch here: https://glitch.com/edit/#!/era-project. The app now connects to my Compose deployment of RethinkDB and pulls the balance for a user when it is given in a form.